+++ HISTORY.rst
@@ -24,6 +24,8 @@ Unreleased
 * ALCN-1030: Add legal mention on fiscal position and display this mention on the invoice report
 * Configure product expiry delays
 * ALCN-1268: Display the zetes state on picking and allow inventory manager to edit this value + search picking with or without an operator for zetes
+* ALCN-1243: Improve claim ticket creation, add link to picking, product, invoice; remove reference, add default values
+
 
 **Bugfixes**
 * Delivery Rounds: Fix delivery round import itinerary